Extract TextInput to component
